requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
deepgram-sdk==3.* # Specify version if needed, 3.* recommended for Agent/latest features
openai
pinecone
psycopg2-binary # Or asyncpg if using async DB operations
firebase-admin # For Authentication
# Add other dependencies like SQLAlchemy/SQLModel later
sqlmodel

# Testing
pytest
pytest-cov
pytest-mock
pytest-asyncio # Added for async test support

# Type checking

# Added for background tasks
fastapi-utils

# Added for TestClient
httpx

# Added for mocking HTTPX requests (optional, useful for external APIs)
respx
